+++ lib/src/default_index/mod.rs
