+++ lib/src/default_revset_graph_iterator.rs
@@ -18,7 +18,7 @@ use std::cmp::{min, Ordering};
 use std::collections::{BTreeMap, HashSet};
 
 use crate::backend::CommitId;
-use crate::default_index_store::{CompositeIndex, IndexEntry, IndexPosition};
+use crate::default_index::{CompositeIndex, IndexEntry, IndexPosition};
 use crate::revset_graph::{RevsetGraphEdge, RevsetGraphEdgeType};
 
 /// Like `RevsetGraphEdge`, but stores `IndexPosition` instead.
