+++ lib/src/lib.rs
@@ -25,7 +25,7 @@ pub mod commit;
 pub mod commit_builder;
 pub mod conflicts;
 pub mod dag_walk;
-pub mod default_index_store;
+pub mod default_index;
 pub mod default_revset_engine;
 pub mod default_revset_graph_iterator;
 pub mod default_submodule_store;
