+++ lib/tests/test_default_revset_graph_iterator.rs
@@ -14,7 +14,7 @@
 
 use itertools::Itertools;
 use jj_lib::commit::Commit;
-use jj_lib::default_index_store::DefaultReadonlyIndex;
+use jj_lib::default_index::DefaultReadonlyIndex;
 use jj_lib::default_revset_engine::{evaluate, RevsetImpl};
 use jj_lib::repo::{ReadonlyRepo, Repo as _};
 use jj_lib::revset::ResolvedExpression;
